requirements.txt
typer
pypdf
python-dotenv
requests
google-genai
pillow
